Kind-of fix #4.
The way SockJS creates the listeners is to go through every listener ever created until one is reached that satisfies the request. As such, it is not possible to remove the listener (as it is a function that does some stuff with regexps to test the URL), or revert back to the listeners prior to the one relating to the channel is added as it may remove future added channels.

So, instead we add another listener to basically give a 404 error.